Fixing bug where mousedown -> focus would cause weird selection behavior for lists.
I don't think it's possible to differentiate between sources of focus events. This is problematic because a mousedown triggers a focus (i.e. enter-from-top on the list), but the subsequent mouseup most likely will generate a click, meaning the highlight will immediately move. This means the list and the recipe both flash and it looks really bad. At the expense of supporting tabbing through the lists, I'm removing the focus event handler. The arrow keys will still work.